% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PresidentialResults2016.R
\name{load2016PresidentialResults}
\alias{load2016PresidentialResults}
\title{Load 2016 presidential election results from the Pettigrew dataverse dataset at Harvard, augmented
by data from some states directly from state sources.}
\usage{
load2016PresidentialResults()
}
\description{
Load 2016 presidential election results from the Pettigrew dataverse dataset at Harvard, augmented
by data from some states directly from state sources.
}

