% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualizations.R
\name{stateDemocraticRepublican2016ResultsChoropleth}
\alias{stateDemocraticRepublican2016ResultsChoropleth}
\title{Produce a red/blue gradient choropleth of the state based upon the percentage of Trump vs Clinton votes}
\usage{
stateDemocraticRepublican2016ResultsChoropleth(state, labels = FALSE)
}
\description{
Produce a red/blue gradient choropleth of the state based upon the percentage of Trump vs Clinton votes
}

