% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualizations.R
\name{stateDemocraticRepublicanChoropleth}
\alias{stateDemocraticRepublicanChoropleth}
\title{Produce a red/blue gradient choropleth of the state}
\usage{
stateDemocraticRepublicanChoropleth(countyLevelDf, state, labels = FALSE,
  stateFIPSColumnName = "State", stateNameColumnName = "StateName",
  stateAbbrColumnName = "StateAbbr", countyFIPSColumnName = "County",
  caption = "", titleFunction = function(stateName) {     "" },
  RDRatioColumnName)
}
\arguments{
\item{state}{the two-letter FIPS abbrevation for the state}

\item{labels}{whether to draw the county names on the map (default is FALSE)}

\item{stateFIPSColumnName}{the name of the column in the provided data frame that contains the numeric FIPS code for the state (default=State)}

\item{stateNameColumnName}{the name of the column in the provided data frame that contains the full state name (default=StateName)}

\item{countyFIPSColumnName}{the name of the column in the provided data frame that contains the 5-digit FIPS code for each county (default=County)}

\item{caption}{a string containing the caption for the map (default is an empty string / no caption)}

\item{titleFunction}{a function, taking the name of the state as a parameter, that returns the string for the title (default is empty string / no title)}

\item{RDRatioColumnName}{the name of the column in the provided data frame that contains the ratio (R/(D+R)) to be displayed on the choropleth}
}
\description{
Produce a red/blue gradient choropleth of the state
}

