% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualizations.R
\name{stateDemocraticRepublicanRegistrationChoropleth}
\alias{stateDemocraticRepublicanRegistrationChoropleth}
\title{Produce a red/blue gradient choropleth of the state}
\usage{
stateDemocraticRepublicanRegistrationChoropleth(state)
}
\arguments{
\item{state}{the two-letter FIPS abbrevation for the state}
}
\description{
Produce a red/blue gradient choropleth of the state
}

