% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualizations.R
\name{stateDemocraticRepublicanRegistrationChoropleth}
\alias{stateDemocraticRepublicanRegistrationChoropleth}
\title{Produce a red/blue gradient choropleth of the state}
\usage{
stateDemocraticRepublicanRegistrationChoropleth(state, labels = FALSE)
}
\arguments{
\item{state}{the two-letter FIPS abbrevation for the state}

\item{labels}{whether to draw the county names on the map (default is FALSE)}
}
\description{
Produce a red/blue gradient choropleth of the state
}

